refactor: Move follow POST/DELETE to follows feature
